+++ AWSCognitoIdentityProvider/AWSCognitoIdentityProviderModel.m
@@ -1414,19 +1414,24 @@ NSString *const AWSCognitoIdentityProviderErrorDomain = @"com.amazonaws.AWSCogni
 
 + (NSValueTransformer *)eventResponseJSONTransformer {
     return [AWSMTLValueTransformer reversibleTransformerWithForwardBlock:^NSNumber *(NSString *value) {
-        if ([value caseInsensitiveCompare:@"Success"] == NSOrderedSame) {
-            return @(AWSCognitoIdentityProviderEventResponseTypeSuccess);
+        if ([value caseInsensitiveCompare:@"Pass"] == NSOrderedSame) {
+            return @(AWSCognitoIdentityProviderEventResponseTypePass);
         }
-        if ([value caseInsensitiveCompare:@"Failure"] == NSOrderedSame) {
-            return @(AWSCognitoIdentityProviderEventResponseTypeFailure);
+        if ([value caseInsensitiveCompare:@"Fail"] == NSOrderedSame) {
+            return @(AWSCognitoIdentityProviderEventResponseTypeFail);
+        }
+        if ([value caseInsensitiveCompare:@"InProgress"] == NSOrderedSame) {
+            return @(AWSCognitoIdentityProviderEventResponseTypeInProgress);
         }
         return @(AWSCognitoIdentityProviderEventResponseTypeUnknown);
     } reverseBlock:^NSString *(NSNumber *value) {
         switch ([value integerValue]) {
-            case AWSCognitoIdentityProviderEventResponseTypeSuccess:
-                return @"Success";
-            case AWSCognitoIdentityProviderEventResponseTypeFailure:
-                return @"Failure";
+            case AWSCognitoIdentityProviderEventResponseTypePass:
+                return @"Pass";
+            case AWSCognitoIdentityProviderEventResponseTypeFail:
+                return @"Fail";
+            case AWSCognitoIdentityProviderEventResponseTypeInProgress:
+                return @"InProgress";
             default:
                 return nil;
         }
@@ -1448,6 +1453,12 @@ NSString *const AWSCognitoIdentityProviderErrorDomain = @"com.amazonaws.AWSCogni
         if ([value caseInsensitiveCompare:@"ForgotPassword"] == NSOrderedSame) {
             return @(AWSCognitoIdentityProviderEventTypeForgotPassword);
         }
+        if ([value caseInsensitiveCompare:@"PasswordChange"] == NSOrderedSame) {
+            return @(AWSCognitoIdentityProviderEventTypePasswordChange);
+        }
+        if ([value caseInsensitiveCompare:@"ResendCode"] == NSOrderedSame) {
+            return @(AWSCognitoIdentityProviderEventTypeResendCode);
+        }
         return @(AWSCognitoIdentityProviderEventTypeUnknown);
     } reverseBlock:^NSString *(NSNumber *value) {
         switch ([value integerValue]) {
@@ -1457,6 +1468,10 @@ NSString *const AWSCognitoIdentityProviderErrorDomain = @"com.amazonaws.AWSCogni
                 return @"SignUp";
             case AWSCognitoIdentityProviderEventTypeForgotPassword:
                 return @"ForgotPassword";
+            case AWSCognitoIdentityProviderEventTypePasswordChange:
+                return @"PasswordChange";
+            case AWSCognitoIdentityProviderEventTypeResendCode:
+                return @"ResendCode";
             default:
                 return nil;
         }
@@ -1991,6 +2006,7 @@ NSString *const AWSCognitoIdentityProviderErrorDomain = @"com.amazonaws.AWSCogni
              @"allowedOAuthFlowsUserPoolClient" : @"AllowedOAuthFlowsUserPoolClient",
              @"allowedOAuthScopes" : @"AllowedOAuthScopes",
              @"analyticsConfiguration" : @"AnalyticsConfiguration",
+             @"authSessionValidity" : @"AuthSessionValidity",
              @"callbackURLs" : @"CallbackURLs",
              @"clientName" : @"ClientName",
              @"defaultRedirectURI" : @"DefaultRedirectURI",
@@ -5612,6 +5628,7 @@ NSString *const AWSCognitoIdentityProviderErrorDomain = @"com.amazonaws.AWSCogni
              @"allowedOAuthFlowsUserPoolClient" : @"AllowedOAuthFlowsUserPoolClient",
              @"allowedOAuthScopes" : @"AllowedOAuthScopes",
              @"analyticsConfiguration" : @"AnalyticsConfiguration",
+             @"authSessionValidity" : @"AuthSessionValidity",
              @"callbackURLs" : @"CallbackURLs",
              @"clientId" : @"ClientId",
              @"clientName" : @"ClientName",
@@ -6018,6 +6035,7 @@ NSString *const AWSCognitoIdentityProviderErrorDomain = @"com.amazonaws.AWSCogni
              @"allowedOAuthFlowsUserPoolClient" : @"AllowedOAuthFlowsUserPoolClient",
              @"allowedOAuthScopes" : @"AllowedOAuthScopes",
              @"analyticsConfiguration" : @"AnalyticsConfiguration",
+             @"authSessionValidity" : @"AuthSessionValidity",
              @"callbackURLs" : @"CallbackURLs",
              @"clientId" : @"ClientId",
              @"clientName" : @"ClientName",